Frontend branch

+++ Backend/src/main/java/com/example/backend/user/UserInterface.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserInterface extends JpaRepository<User, Integer> {
-    Optional<User> findUserByUserId(Integer userId);
+    Optional<User> findUserByUserId(Long userId);
     Optional<User> findUserByEmail(String email);
     Optional<User> findUserByUsername(String username);
